fix: address leave policy PR review comments
- fold the balance calculation back into PolicyLeaveServiceImpl instead of a
  separate PolicyLeaveBalanceService
- credit a future leave cycle in full so next-year cards no longer show 0
- skip the second accrual walk when creditedUpTo already covers the cycle
- drop the dead employeeLeavePolicyList mapper method
- keep the accrual preview's running total unrounded so it matches what the
  backend credits
- gate the join-date warning on the join-date option, drop the never-rendered
  disabled-button title
- keep both effective-date radios direct children of the radiogroup and scroll
  only the accrual table so the datepicker popper is not clipped
- move the table header casing to CSS, restore the page heading on the legacy
  leave-types tab, and read isUnlimited/formatDays on the policy card
