+++ backend/src/main/java/com/skapp/community/leaveplanner/service/PolicyLeaveService.java
@@ -1,6 +1,8 @@
 package com.skapp.community.leaveplanner.service;
 
 import com.skapp.community.common.payload.response.ResponseEntityDto;
+import com.skapp.community.leaveplanner.model.EmployeeLeavePolicy;
+import com.skapp.community.leaveplanner.payload.PolicyLeaveBalanceDto;
 import com.skapp.community.leaveplanner.payload.request.PolicyLeaveAvailabilityRequestDto;
 import com.skapp.community.leaveplanner.payload.request.PolicyLeaveRequestDto;
 import com.skapp.community.leaveplanner.payload.request.PolicyLeaveRequestFilterDto;
@@ -25,5 +27,12 @@ public interface PolicyLeaveService {
 	ResponseEntityDto nudgePolicyLeaveRequestManagers(Long id);
 
 	ResponseEntityDto getPolicyLeaveRequestNudgeStatus(Long id);
+
+	/**
+	 * Derives the leave balance of a managed assignment for the given leave cycle. The
+	 * caller must supply an assignment that is still attached to an open persistence
+	 * context.
+	 */
+	PolicyLeaveBalanceDto calculateBalanceForYear(EmployeeLeavePolicy assignment, int year);
 
 }
