+++ frontend/src/community/leave/components/molecules/UserLeavePolicies/LeavePolicyCard.tsx
@@ -3,10 +3,7 @@ import { FC } from "react";
 
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { formatDays, getEmoji } from "~community/common/utils/commonUtil";
-import {
-  EmployeeLeavePolicyType,
-  PolicyType
-} from "~community/leave/types/LeavePolicyTypes";
+import { EmployeeLeavePolicyType } from "~community/leave/types/LeavePolicyTypes";
 
 interface Props {
   policy: EmployeeLeavePolicyType;
@@ -25,10 +22,10 @@ const LeavePolicyCard: FC<Props> = ({
 }) => {
   const translateText = useTranslator("leaveModule", "leavePolicyAssignment");
 
-  const isFlexiblePolicy = policy.policyType === PolicyType.FLEXIBLE;
+  const isFlexiblePolicy = policy.isUnlimited;
 
   return (
-    <Card className="flex h-[96px] flex-row items-center justify-between gap-4 bg-white p-6!">
+    <Card className="flex min-h-[96px] flex-row items-center justify-between gap-4 bg-white p-6!">
       <div className="flex min-w-0 flex-row items-center gap-6">
         {isFlexiblePolicy ? (
           <span
@@ -42,8 +39,8 @@ const LeavePolicyCard: FC<Props> = ({
           <div
             className="flex shrink-0 items-baseline gap-0.5"
             aria-label={translateText(["leavesRemainingLabel"], {
-              remaining: policy.balanceInDays,
-              total: policy.totalDaysAllocated
+              remaining: formatDays(policy.balanceInDays),
+              total: formatDays(policy.totalDaysAllocated)
             })}
           >
             <span className="h1b text-black">
